Remove blog article tags
