+++ content/blog/dispatch.mdx
@@ -2,7 +2,6 @@
 title: "10x Your Claude Code Window Size with Dispatch"
 description: "Your Claude Code session has a fixed context window. Dispatch multiplies it by turning your session into a lightweight orchestrator while dedicated workers each run with their own full context."
 thumbnail: /images/blog/dispatch-before-after.jpg
-tags: ["AI", "Claude Code", "Developer Tools"]
 published: "2026-02-23T07:44:46-08:00"
 modified: "2026-02-24T23:46:01-08:00"
 ---
