+++ content/blog/game-factory.mdx
@@ -3,7 +3,6 @@ title: "An Autonomous Game Factory with Claude Code and a $24 VPS"
 description: "Anyone can prompt 'make a flappy bird game.' Making hundreds of unique, styled,
 working games fully autonomously is a different problem entirely."
 thumbnail: /images/blog/game-factory-grid-5x5.png
-tags: ["AI", "Claude Code", "Autonomous Agents", "Game Development"]
 published: "2026-04-01T00:19:02-07:00"
 modified: "2026-04-01T00:42:57-07:00"
 ---
