+++ content/blog/invariant-engineering.mdx
@@ -2,7 +2,6 @@
 title: "Invariant Engineering: Why Your AI Agent Is Either Broken or Boring"
 description: "Deciding what your AI system is and isn't allowed to vary. Most teams get this wrong in one of two directions."
 thumbnail: /images/blog/invariant-engineering.png
-tags: ["AI", "Agentic Engineering", "Developer Tools"]
 published: "2026-04-12T18:01:52-07:00"
 modified: "2026-04-12T22:39:29-07:00"
 ---
