+++ content/blog/levels-of-agentic-engineering.mdx
@@ -2,7 +2,6 @@
 title: "The 8 Levels of Agentic Engineering"
 description: "AI's coding ability is outpacing our ability to wield it effectively. That gap closes in levels — 8 of them. Here's the progression from tab complete to autonomous agent teams."
 thumbnail: /images/blog/agentic_thumbnail.jpg
-tags: ["AI", "Agentic Engineering", "Developer Tools"]
 published: "2026-03-10T01:31:01-07:00"
 modified: "2026-03-11T22:06:21-07:00"
 ---
