+++ content/blog/tokenmaxxing.mdx
@@ -1,7 +1,6 @@
 ---
 title: "How to Tokenmaxx"
 description: "Tokenmaxxing is fine if the tokens go toward planning, verification, context, feedback loops, and agent-friendly output. It is a terrible proxy for AI competence on its own."
-tags: ["AI", "LLMs", "Work"]
 published: "2026-05-14T02:00:37-07:00"
 modified: "2026-05-20T00:00:00-07:00"
 ---
