+++ next.config.mjs
@@ -4,12 +4,22 @@ const nextConfig = {
     return [
       {
         source: "/blog/traits-of-ai-native-people",
-        destination: "/blog/token-count-is-a-terrible-proxy-for-ai-competence",
+        destination: "/tokenmaxxing",
         permanent: true,
       },
       {
         source: "/blog/token-count-is-a-bad-proxy-for-ai-native-work",
-        destination: "/blog/token-count-is-a-terrible-proxy-for-ai-competence",
+        destination: "/tokenmaxxing",
+        permanent: true,
+      },
+      {
+        source: "/blog/token-count-is-a-terrible-proxy-for-ai-competence",
+        destination: "/tokenmaxxing",
+        permanent: true,
+      },
+      {
+        source: "/blog/tokenmaxxing",
+        destination: "/tokenmaxxing",
         permanent: true,
       },
     ];
